提交 修改 保存时 的主键 类型

+++ 9iuspring/src/org/iu9/frame/dao/BaseJdbcDaoImpl.java
@@ -476,7 +476,7 @@ public abstract class BaseJdbcDaoImpl extends BaseLogger implements
 			} else {
 				getJdbc().update(sql.toString(), ss, keyHolder);
 			}
-			return keyHolder.getKey();
+			return keyHolder.getKey().longValue();
 		}
 	}
